changed constructor from int angle to float angle for gameobject & extended class

+++ src/tankrotationexample/game/Bullet.java
@@ -8,7 +8,7 @@ public abstract class Bullet extends Moving{
     private final int R = 7;
     private Rectangle hitBox;
 
-    Bullet(int currX, int currY, int currVX, int currVY, int currAngle, BufferedImage currImg){
+    Bullet(int currX, int currY, int currVX, int currVY, float currAngle, BufferedImage currImg){
         super(currX, currY, currVX, currVY, currAngle, currImg);
     }
