+++ src/tankrotationexample/game/MachineGunMode.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 
 public class MachineGunMode extends Powerup{
 
-    MachineGunMode(int currX, int currY, int currVX, int currVY, int currAngle, BufferedImage currImg){
+    MachineGunMode(int currX, int currY, int currVX, int currVY, float currAngle, BufferedImage currImg){
         super(currX, currY, currVX, currVY, currAngle, currImg);
     }
