+++ src/tankrotationexample/game/Powerup.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 
 public abstract class Powerup extends Moving{
 
-    Powerup(int currX, int currY, int currVX, int currVY, int currAngle, BufferedImage currImg){
+    Powerup(int currX, int currY, int currVX, int currVY, float currAngle, BufferedImage currImg){
         super(currX, currY, currVX, currVY, currAngle, currImg);
     }
