+++ src/tankrotationexample/game/Wall.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 
 public abstract class Wall extends Stationary{
 
-    Wall(int currX, int currY, int currAngle, BufferedImage currImage){
+    Wall(int currX, int currY, float currAngle, BufferedImage currImage){
         super(currX, currY, currAngle, currImage);
     }
